update
quan ly tai khoan

+++ Shop_Mobi/Databases/tbl_sanpham.cs
@@ -22,7 +22,7 @@ namespace Shop_Mobi.Databases
         [StringLength(150)]
         public string ten { get; set; }
 
-        public decimal? gia { get; set; }
+        public int? gia { get; set; }
 
         public int? soluong { get; set; }
